+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ops/CustomOp.java
@@ -31,14 +31,39 @@ public interface CustomOp {
     boolean isInplaceCall();
 
 
+    /**
+     * Input arguments for
+     * this op
+     * @return
+     */
     List<INDArray> getInputArguments();
 
+    /**
+     * Output arguments for this
+     * @return
+     */
     List<INDArray> getOutputArguments();
 
+    /**
+     * Integer input arguments for this
+     * @return
+     */
     List<Integer> getIArguments();
 
+    /**
+     * Floating point arguments
+     * for this op.
+     * The "T" stands for
+     * a generic floating point type
+     * in c++.
+     * @return
+     */
     List<Double> getTArguments();
 
 
+    /**
+     * Calculate the output shape for this op
+     * @return
+     */
     List<int[]> calculateOutputShape();
 }
