+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ops/Module.java
@@ -1,0 +1,20 @@
+package org.nd4j.linalg.api.ops;
+
+/**
+ * A Module is a {@link CustomOp}
+ * with varying input arguments
+ * and automatically calculated outputs
+ * defined at a higher level than c++.
+ *
+ * A Module is meant to be a way of implementing custom operations
+ * in straight java/nd4j.
+ */
+public interface Module extends CustomOp {
+
+    /**
+     *
+     */
+    void exec();
+
+
+}
