+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ops/impl/layers/convolution/config/Conv2DConfig.java
@@ -1,4 +1,4 @@
-package org.nd4j.linalg.api.ops.impl.transforms.convolution.config;
+package org.nd4j.linalg.api.ops.impl.layers.convolution.config;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
